requirements.txt
autopep8==1.5
boto3==1.14.1
botocore==1.17.1
captcha==0.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-discover-runner==1.0
django-email==0.1.10
django-environ==0.4.5
djangorestframework==3.10.0
environ==1.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.4
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
azure-storage-file-datalake==12.1.1
azure-storage-queue==12.1.2
azure-common==1.1.25 
azure-storage-common==2.1.0 
azure-storage-file==2.1.0
azure-core==1.8.0 
azure-storage-blob==12.4.0 